Code coverage improved with more regression tests. Also, example input and submission script included for CI tests
