+++ examples/data/elasticity_tensor_workflow/defor_1/geometry.in
@@ -1,7 +1,7 @@
 #===============================================================================
 # Created using the Atomic Simulation Environment (ASE)
 
-# Thu Mar 12 15:05:14 2026
+# Thu Mar 12 16:46:52 2026
 
 #=======================================================
 lattice_vector 3.5561438210035812 -0.0000000019583704 0.0000000009889083
