+++ examples/data/elasticity_tensor_workflow/defor_1/input.py
@@ -1,0 +1,6 @@
+'''
+Test input file CI tests. Replace this with the input file for the calculator of your choice
+'''
+print('Test input file CI tests. Replace this with the input file for the calculator of your choice')
+
+
