+++ examples/data/elasticity_tensor_workflow/defor_3/geometry.in
@@ -1,7 +1,7 @@
 #===============================================================================
 # Created using the Atomic Simulation Environment (ASE)
 
-# Thu Mar 12 15:05:14 2026
+# Thu Mar 12 16:46:52 2026
 
 #=======================================================
 lattice_vector 3.5211070668383782 -0.0000000019778571 0.0000000009889083
